The latency toxic adds a delay to all data going through the proxy. The delay is equal to latency +/- jitter.

The latency toxic adds a delay to all data going through the proxy. The delay is equal to latency +/- jitter.

The bandwidth toxic limits a connection to a maximum number of kilobytes per second.

The slow_close toxic delays the TCP socket from closing until delay has elapsed.

The timeout toxic stops all data from getting through, and closes the connection after timeout. If timeout is 0, the connection won't close, and data will be delayed until the toxic is removed.

The slicer toxic slices TCP data up into small bits, optionally adding a delay between each sliced "packet".

The limit_data toxic closes the connection when transmitted data exceeded limit.

latency(latency, jitter = 0)

bandwidth(rate)

slow_close(delay)

timeout(timeout)

slicer(average_size, size_variation, delay = 0)

limit_data(bytes)

Arguments

latency
Latency in milliseconds

jitter
Time in milliseconds

rate
Rate in KB per second

delay
Time in milliseconds

timeout
Time in milliseconds

average_size
Size in bytes of an average packet

size_variation
Variation in bytes of an average packet (should be smaller than average_size)

bytes
The number of bytes it should transmit before connection is closed